refactor(cli): rename llm-warning module to llm-failure
The module models a failure (LlmFailure, detectLlmFailure,
formatLlmFailureError) that exits 8, not a warning. Rename the file +
test to match the symbols and the LLM_FAILURE exit code, and repoint the
imports in score.ts and the test.

Refs #142
