+++ packages/cli/src/llm-failure.ts
